md_dapt_pipeline/pyproject.toml
[project]
name = "md-dapt-pipeline"
version = "0.1.0"
description = "Markdown conversational DAPT pipeline"
requires-python = ">=3.10"

dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.pyright]
include = ["."]
extraPaths = ["."]
